Migrated from net server to http server
- Added new mime type ico for favicon
- Deleted request.js and response.js
- Created queryParser.js for parsing query.
- Changes in responseHandler.js
-> Changed function name of generateDefaultResponse to error404; this
name is more precise
-> Delete generateResponse => Very less usage, because with http
server you can send response in one or two lines.
-> Updated saveComment -> Now it will take queryText and parse it
itself.
-> Deleted formatCommentsToHTML and addCommentsToHtml, and put them
into addCommentsToTemplate, because both are doing small work which
addCommentsToTemplate can do itself.
-> Added function serveTemplate =>  It will serve templates.
-> Refactored guestBookHandler => it is just serving the template
after saving the comment.
-> Added the index page(/) logic in serverStaticPage function itself.
-> Changed processRequest to handleRequest => using a variable to save
the handler reference and later called it.

+++ server.js
@@ -1,23 +1,11 @@
-const { Server } = require('net');
-const { processRequest, dataStorePath } = require('./lib/responseHandler');
-const { writeFileSync } = require('fs');
+const http = require('http');
+const { handleRequest, dataStorePath } = require('./lib/responseHandler');
+const { writeFileSync, existsSync } = require('fs');
 
-const server = new Server();
+const server = http.Server(handleRequest);
 const PORT = 8000;
 
 server.listen(process.argv[2] || PORT);
 
-server.on('listening', () => {
-  writeFileSync(dataStorePath, '[]');
-  const { address, family, port } = server.address();
-  console.log(`\nListening on port ${address}:${port} with ${family}\n`);
-});
-
-server.on('connection', socket => {
-  socket.setEncoding('utf8');
-  socket.on('data', data => {
-    const response = processRequest(data);
-    response.send(socket);
-    socket.end();
-  });
-});
+const dataStoreExists = existsSync(dataStorePath);
+if (!dataStoreExists) writeFileSync(dataStorePath, '[]');
